+++ src/main/java/absolutelyaya/ultracraft/registry/ItemRegistry.java
@@ -103,6 +103,8 @@ public class ItemRegistry
 			new Identifier(Ultracraft.MOD_ID, "soap"), new SoapItem(new FabricItemSettings().maxCount(4).rarity(Rarity.EPIC)));
 	public static final AttractorNailgunItem ATTRACTOR_NAILGUN = Registry.register(Registries.ITEM,
 			new Identifier(Ultracraft.MOD_ID, "attractor_nailgun"), new AttractorNailgunItem(new FabricItemSettings().maxCount(1)));
+	public static final OverheatNailgunItem OVERHEAT_NAILGUN = Registry.register(Registries.ITEM,
+			new Identifier(Ultracraft.MOD_ID, "overheat_nailgun"), new OverheatNailgunItem(new FabricItemSettings().maxCount(1)));
 	
 	//Spawn Eggs
 	public static final SpawnEggItem FILTH_SPAWN_EGG = Registry.register(Registries.ITEM,
@@ -233,6 +235,7 @@ public class ItemRegistry
 			content.add(CORE_SHOTGUN);
 			content.add(PUMP_SHOTGUN);
 			content.add(ATTRACTOR_NAILGUN);
+			content.add(OVERHEAT_NAILGUN);
 			content.add(MACHINE_SWORD.getDefaultStack(MachineSwordItem.Type.NORMAL));
 			content.add(MACHINE_SWORD.getDefaultStack(MachineSwordItem.Type.TUNDRA));
 			content.add(MACHINE_SWORD.getDefaultStack(MachineSwordItem.Type.AGONY));