+++ changelog.md
@@ -1,5 +1,9 @@
 ## Major Changes
 - Added Minecart Parries
+- Added Overheat Nailgun
+  - Primary fire of all Nailgun variants now builds up "heat"
+  - The higher the heat, the slower the fire rate (this one only applies to the Overheat Nailgun)
+  - Alt fire uses a "heatsink", firing 5 nails and consuming 5 "heat" per tick until heat is used up.
 ## Settings & Gamerules
 ## Commands
 ## Tweaks